add: Added anti link event
